Moved generate_password function to common
