+++ edx_django_utils/common/__init__.py
@@ -1,0 +1,20 @@
+"""
+The intent of this file is that all importable public code will live here.
+"""
+
+import random
+import string
+
+
+def generate_password(length=12, chars=string.ascii_letters + string.digits):
+    """Generate a valid random password"""
+    if length < 8:
+        raise ValueError("password must be at least 8 characters")
+
+    choice = random.SystemRandom().choice
+
+    password = ''
+    password += choice(string.digits)
+    password += choice(string.ascii_letters)
+    password += ''.join([choice(chars) for _i in range(length - 2)])
+    return password
